pyproject.toml
[tool.poetry]
name = "groundhog"
version = "0.1.0"
description = ""
authors = ["Joseph Bettenhausen <21023218+jbettenh@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
flask = "^2.2.2"
flask-session = "^0.4.0"
psycopg2 = "^2.9.5"
flask-sqlalchemy = "^3.0.2"
requests = "^2.28.1"
python-dotenv = "^0.21.0"
folium = "^0.13.0"
geopy = "^2.3.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
black = "^22.10.0"
pytest-cov = "^4.0.0"
testing-postgresql = "^1.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
